requirements.txt
PyGithub
python-dotenv
sentence-transformers
scikit-learn
ollama